Move hooks into their own custom hooks folder
