+++ frontend_react/src/api/SocketService.ts
@@ -1,7 +1,7 @@
 import { SocketServiceInterface } from "../utils/SocketServiceInterface";
 import { User } from "../utils/User";
-import { addLiveUser, loading, setAlert } from "../redux/actions";
-import store from "../redux/store";
+import { addLiveUser, loading, setAlert } from "../state/actions";
+import store from "../state/store";
 
 let retries = 0;
 const timeout = 60000;
