+++ frontend_react/src/components/Map/Map.tsx
@@ -9,10 +9,11 @@ import { useSelector } from "react-redux";
 import { AppState } from "../../utils/AppState";
 import { Theme, useMediaQuery } from "@material-ui/core";
 import { MapBounds } from "../../utils/MapBounds";
+import useProcessedUsers from "../../hooks/useMapMarkers";
 
 type MapProps = {
   onMapBoundsChange: (mapBounds?: MapBounds) => void;
-}
+};
 
 const defaultCenter = { lat: 48.354594, lng: -99.99805 };
 
@@ -29,6 +30,34 @@ const Map = (props: MapProps) => {
     (state: AppState) => state.historicalUsers
   );
   const newUser: User | null = useSelector((state: AppState) => state.newUser);
+  // Set historical markers when historical users prop changes
+  const processedUsers = useProcessedUsers(liveUsers, historicalUsers);
+
+  useEffect(() => {
+    if (!_.isNil(processedUsers)) {
+      const markers = processedUsers?.map((user, index) => {
+        const open = newUser && _.isEqual(user, newUser) ? true : false;
+        return (
+          <MapMarker
+            key={index}
+            user={user}
+            open={open}
+            lat={user.payload.location.latitude}
+            lng={user.payload.location.longitude}
+            onMarkerClick={handleMarkerClick}
+          ></MapMarker>
+        );
+      });
+      setMarkers(markers);
+
+      if (_.isNil(historicalUsers) && !_.isNil(newUser)) {
+        setCenter({
+          lat: newUser.payload.location.latitude,
+          lng: newUser.payload.location.longitude,
+        });
+      }
+    }
+  }, [processedUsers]);
 
   // Hide map control for mobile screens
   const showMapControl = useMediaQuery((theme: Theme) =>
@@ -37,7 +66,7 @@ const Map = (props: MapProps) => {
   const defaultMapOptions = (maps: Maps) => {
     return {
       zoomControl: false,
-      minZoom: 2.5,
+      minZoom: 3,
       restriction: {
         latLngBounds: { north: 85, south: -85, west: -180, east: 180 },
       },
@@ -55,77 +84,9 @@ const Map = (props: MapProps) => {
     };
   };
 
-  // Set historical markers when historical users prop changes
-  useEffect(() => {
-    // Group a user list by location
-    const groupByLocation = (users: any[]): { [location: string]: any } => {
-      return users.reduce((storage, user) => {
-        // Create a string key based on combination of lat/long
-        const group = `${user.payload.location.latitude}_${user.payload.location.longitude}`;
-        storage[group] = storage[group] || [];
-        storage[group].push(user);
-        return storage;
-      }, {});
-    };
-
-    const updateMarkers = (users: User[], newUser: User | null) => {
-      const processedUsers: User[] = [];
-
-      // Group users by location
-      const groupedUsers = groupByLocation(users);
-
-      // For each location keep the latest user by date
-      for (const key in groupedUsers) {
-        const users: User[] = groupedUsers[key];
-        let latestUser: User = users[0];
-
-        users.forEach((user) => {
-          if (!_.isNil(user.date) && !_.isNil(latestUser.date)) {
-            latestUser = user.date > latestUser.date ? user : latestUser;
-          }
-        });
-
-        processedUsers.push(latestUser);
-      }
-
-      // Sort in descending order by latitude to avoid overlapping on map
-      processedUsers.sort(
-        (a: User, b: User) =>
-          b.payload.location.latitude - a.payload.location.latitude
-      );
-
-      return processedUsers.map((user, index) => {
-        const open = newUser && _.isEqual(user, newUser) ? true : false;
-        return (
-          <MapMarker
-            key={index}
-            user={user}
-            open={open}
-            lat={user.payload.location.latitude}
-            lng={user.payload.location.longitude}
-            onMarkerClick={handleMarkerClick}
-          ></MapMarker>
-        );
-      });
-    };
-
-    let markers: ReactElement[];
-    if (_.isNil(historicalUsers)) {
-      markers = updateMarkers(liveUsers, newUser);
-      if (!_.isNil(newUser)) {
-        setCenter({
-          lat: newUser.payload.location.latitude,
-          lng: newUser.payload.location.longitude,
-        });
-      }
-    } else {
-      markers = updateMarkers(historicalUsers, null);
-    }
-
-    setMarkers(markers);
-  }, [liveUsers, historicalUsers]);
-
-  const getMapBounds = (bounds: google.maps.LatLngBounds | undefined): MapBounds | undefined => {
+  const getMapBounds = (
+    bounds: google.maps.LatLngBounds | undefined
+  ): MapBounds | undefined => {
     if (_.isNil(bounds)) {
       return undefined;
     }
